fix code quality and storage bug
